Add static constants for role ids, improve users and roles tests and add some tests

+++ hivelvet-backend/app/src/Actions/Roles/Delete.php
@@ -24,6 +24,7 @@ namespace Actions\Roles;
 
 use Actions\Delete as DeleteAction;
 use Actions\RequirePrivilegeTrait;
+use Enum\ResponseCode;
 use Models\Role;
 
 /**
@@ -38,21 +39,25 @@ class Delete extends DeleteAction
         $role    = new Role();
         $role_id = $params['id'];
         $role->load(['id = ?', $role_id]);
-        $nbUsers = $role->getRoleUsers();
+        if ($role->valid()) {
+            $nbUsers = $role->getRoleUsers();
 
-        // delete users and permissions
-        $result = $role->deleteUsersAndPermissions();
+            // delete users and permissions
+            $result = $role->deleteUsersAndPermissions();
 
-        if ($result) {
-            // delete role after deleting assigned users and permissions
-            parent::execute($f3, $params);
-            // if role have users assigned return lecturer role to get switched users
-            if ($nbUsers > 0) {
-                $result = $role->getLecturerRole();
-                $this->renderJson(['lecturer' => $result]);
-            } else {
-                $this->renderJson(['result' => 'success']);
+            if ($result) {
+                // delete role after deleting assigned users and permissions
+                parent::execute($f3, $params);
+                // if role have users assigned return lecturer role to get switched users
+                if ($nbUsers > 0) {
+                    $result = $role->getLecturerRole();
+                    $this->renderJson(['lecturer' => $result]);
+                } else {
+                    $this->renderJson(['result' => 'success']);
+                }
             }
+        } else {
+            $this->renderJson([], ResponseCode::HTTP_NOT_FOUND);
         }
     }
 }
